Fixup grMaterial_801C9604 usage and some pointer math in gr
