+++ src/melee/gr/grcorneria.c
@@ -67,7 +67,7 @@ struct grCorneria_YakumonoParam {
     /* 0x78 */ s32 x78;
     /* 0x7C */ s32 x7C;
     /* 0x80 */ s32 x80;
-    /* 0x84 */ s32 x84;
+    /* 0x84 */ void* x84;
     /* 0x88 */ f32 x88;
 };
