+++ src/melee/gr/grkongo.c
@@ -5,7 +5,6 @@
 #include <math.h>
 
 #include "granime.h"
-#include "grkongo.static.h"
 #include "grmaterial.h"
 #include "ground.h"
 #include "grzakogenerator.h"
