+++ src/melee/gr/grlast.c
@@ -90,8 +90,12 @@ Vec3 const grLast_803B8480 = { 1.0f, 1.0f, 1.0f };
 Vec3 const grLast_803B848C = { 0.0f, 1.0f, 0.0f };
 Vec3 const grLast_803B8498 = { 0.0f, 0.0f, 1.0f };
 
-/// @todo yakumono struct
-static int* yakumono_param;
+static struct grLast_YakumonoParam {
+    void* x0;
+    void* x4;
+    void* x8;
+    void* xC;
+}* yakumono_param;
 
 static void grLast_OnDemoInit(enum_t arg0)
 {
@@ -585,7 +589,8 @@ static void grLast_8021B2E8(Ground_GObj* gobj)
 
     gp = GET_GROUND(gobj);
     if (Stage_80225194() == 0xB0 || // Boss fight vs Master Hand
-        Stage_80225194() == 0xFB) { // Event match #50
+        Stage_80225194() == 0xFB)
+    { // Event match #50
         temp_f1 = gm_8016ECE8();
         switch (gp->u.map.xC4_b2_25) {
         case 1:
@@ -898,7 +903,7 @@ static void grLast_8021B920(Ground_GObj* gobj_, int arg1)
     case 13:
         grLast_8021C40C(gobj, grNLa_804DBBD8, 120.0F);
         for (i = 0; i < 5; i++) {
-            grMaterial_801C9604(gp->u.map.lv_gobj[i], yakumono_param[0], 0);
+            grMaterial_801C9604(gp->u.map.lv_gobj[i], yakumono_param->x0, 0);
             grMaterial_801C9698(gp->u.map.lv_gobj[i]);
         }
         gp->u.map.xC4_b26 = true;
@@ -913,7 +918,7 @@ static void grLast_8021B920(Ground_GObj* gobj_, int arg1)
         HSD_ASSERT(0x4D2, gp->u.map.lv_gobj[5]);
         HSD_GObjGXLink_803909D8(gp->u.map.lv_gobj[5], gobj);
         do_anime(gp->u.map.lv_gobj[5], 5, 0);
-        grMaterial_801C9604(gp->u.map.lv_gobj[5], yakumono_param[1], 0);
+        grMaterial_801C9604(gp->u.map.lv_gobj[5], yakumono_param->x4, 0);
         grMaterial_801C9698(gp->u.map.lv_gobj[5]);
         Ground_801C1E00(0);
         gp->u.map.xC4_b26 = true;
@@ -924,7 +929,7 @@ static void grLast_8021B920(Ground_GObj* gobj_, int arg1)
         break;
     case 16:
         grLast_8021C40C(gobj, grNLa_804DBBE0, 60.0F);
-        grMaterial_801C9604(gp->u.map.lv_gobj[5], yakumono_param[2], 0);
+        grMaterial_801C9604(gp->u.map.lv_gobj[5], yakumono_param->x8, 0);
         grMaterial_801C9698(gp->u.map.lv_gobj[5]);
         gp->u.map.xC4_b26 = true;
         break;
@@ -938,7 +943,7 @@ static void grLast_8021B920(Ground_GObj* gobj_, int arg1)
                                  grNLa_803E8010[i][3], 0.0F, 1.0F);
             }
             do_anime(gp->u.map.lv_gobj[i], i, 0);
-            grMaterial_801C9604(gp->u.map.lv_gobj[i], yakumono_param[3], 0);
+            grMaterial_801C9604(gp->u.map.lv_gobj[i], yakumono_param->xC, 0);
             grMaterial_801C9698(gp->u.map.lv_gobj[i]);
         }
