+++ src/melee/gr/grmaterial.h
@@ -40,7 +40,7 @@ grMaterial_801C8CFC(int, int, Ground*, HSD_JObj*,
 /* 1C9490 */ void grMaterial_801C9490(Item_GObj* gobj, CommandInfo* cmd);
 /* 1C94D8 */ void grMaterial_801C94D8(void*);
 /* 1C95C4 */ void grMaterial_801C95C4(HSD_GObj*);
-/* 1C9604 */ void grMaterial_801C9604(HSD_GObj* bg, int, bool);
+/* 1C9604 */ void grMaterial_801C9604(HSD_GObj* bg, void*, bool);
 /* 1C9664 */ void fn_801C9664(Item_GObj* gobj, CommandInfo* cmd, int arg2);
 /* 1C9698 */ void grMaterial_801C9698(HSD_GObj*);
